requirements.txt
discord.py[voice]
yt-dlp
fastapi
uvicorn[standard]
websockets
requests
python-dotenv
aiofiles
aiohttp
